feat: 설정 위치 주변 500m 내 운영중인 매장 조회 기능 추가 (고객 기능), mybatis 사용

+++ src/main/java/com/zerobase/SelfWash/administer_store/domain/dto/StoreDto.java
@@ -19,8 +19,8 @@ public class StoreDto {
   private long storeId;
   private long ownerId;
   private String address;
-  private float latitude;
-  private float longitude;
+  private double latitude;
+  private double longitude;
   private String notes;
   private boolean opened;
   private boolean approved;
